+++ confignode/src/main/java/org/apache/iotdb/confignode/manager/ProcedureManager.java
@@ -361,8 +361,8 @@ public class ProcedureManager {
     final CreateTriggerProcedure createTriggerProcedure =
         new CreateTriggerProcedure(triggerInformation, jarFile);
     try {
-      final int planSize = new UpdateProcedurePlan(createTriggerProcedure).getSerializedSize();
-      if (planSize > planSizeLimit) {
+      if (jarFile != null
+          && new UpdateProcedurePlan(createTriggerProcedure).getSerializedSize() > planSizeLimit) {
         return new TSStatus(TSStatusCode.CREATE_TRIGGER_ERROR.getStatusCode())
             .setMessage(
                 String.format(
