+++ confignode/src/main/java/org/apache/iotdb/confignode/manager/UDFManager.java
@@ -21,10 +21,12 @@ package org.apache.iotdb.confignode.manager;
 
 import org.apache.iotdb.common.rpc.thrift.TDataNodeLocation;
 import org.apache.iotdb.common.rpc.thrift.TSStatus;
+import org.apache.iotdb.commons.conf.IoTDBConstant;
 import org.apache.iotdb.commons.udf.UDFInformation;
 import org.apache.iotdb.confignode.client.DataNodeRequestType;
 import org.apache.iotdb.confignode.client.async.AsyncDataNodeClientPool;
 import org.apache.iotdb.confignode.client.async.handlers.AsyncClientHandler;
+import org.apache.iotdb.confignode.conf.ConfigNodeDescriptor;
 import org.apache.iotdb.confignode.consensus.request.read.function.GetFunctionTablePlan;
 import org.apache.iotdb.confignode.consensus.request.read.udf.GetUDFJarPlan;
 import org.apache.iotdb.confignode.consensus.request.write.function.CreateFunctionPlan;
@@ -56,6 +58,12 @@ public class UDFManager {
 
   private final ConfigManager configManager;
   private final UDFInfo udfInfo;
+
+  private final long planSizeLimit =
+      ConfigNodeDescriptor.getInstance()
+              .getConf()
+              .getPartitionRegionRatisConsensusLogAppenderBufferSize()
+          - IoTDBConstant.RAFT_LOG_BASIC_SIZE;
 
   public UDFManager(ConfigManager configManager, UDFInfo udfInfo) {
     this.configManager = configManager;
@@ -89,16 +97,22 @@ public class UDFManager {
       }
 
       final boolean needToSaveJar = isUsingURI && udfInfo.needToSaveJar(jarName);
+      CreateFunctionPlan createFunctionPlan =
+          new CreateFunctionPlan(udfInformation, needToSaveJar ? new Binary(jarFile) : null);
+      if (needToSaveJar && createFunctionPlan.getSerializedSize() > planSizeLimit) {
+        return new TSStatus(TSStatusCode.CREATE_TRIGGER_ERROR.getStatusCode())
+            .setMessage(
+                String.format(
+                    "Fail to create UDF[%s], the size of Jar is too large, you can increase the value of property 'partition_region_ratis_log_appender_buffer_size_max' on ConfigNode",
+                    udfName));
+      }
 
       LOGGER.info(
           "Start to add UDF [{}] in UDF_Table on Config Nodes, needToSaveJar[{}]",
           udfName,
           needToSaveJar);
 
-      return configManager
-          .getConsensusManager()
-          .write(new CreateFunctionPlan(udfInformation, needToSaveJar ? new Binary(jarFile) : null))
-          .getStatus();
+      return configManager.getConsensusManager().write(createFunctionPlan).getStatus();
     } catch (Exception e) {
       LOGGER.warn(e.getMessage(), e);
       return new TSStatus(TSStatusCode.EXECUTE_STATEMENT_ERROR.getStatusCode())
